# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Plutonium SMP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: Get user's tickets
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "200":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketWithMessages"
  /tickets/{id}/messages:
    post:
      operationId: sendTicketMessage
      tags: [tickets]
      summary: Send a message in a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
  /tickets/{id}/close:
    post:
      operationId: closeTicket
      tags: [tickets]
      summary: Close a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        subject:
          type: string
        category:
          type: string
          enum: [support, purchase_issues, report_player, general]
        status:
          type: string
          enum: [open, pending, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - subject
        - category
        - status
        - priority
        - messageCount
        - createdAt
        - updatedAt
    TicketMessage:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        userId:
          type: string
        username:
          type: string
        role:
          type: string
        content:
          type: string
        isStaff:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - ticketId
        - userId
        - username
        - content
        - isStaff
        - createdAt
    TicketWithMessages:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/Ticket"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TicketMessage"
      required:
        - ticket
        - messages
    CreateTicketRequest:
      type: object
      properties:
        subject:
          type: string
        category:
          type: string
        priority:
          type: string
        message:
          type: string
      required:
        - subject
        - category
        - message
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...